requirements.txt
beautifulsoup4==4.3.2
bleach==1.4.1
chardet==2.3.0
csvImporter==0.1.3.2
diff-match-patch==20121119
Django==1.7.5
django-admin-jqueryui==1.8.22
django-admin-sortable==1.8.2
django-appconf==1.0.1
django-ckeditor==4.4.7
django-contrib-comments==1.5
django-csvimport==2.4
django-email-obfuscator==0.1.4
django-extensions==1.5.2
django-extlog==0.8
django-image-cropping==1.0.1
django-import-export==0.2.6
django-inline-ordering==1.0.2
django-nested-inline==0.3.4
django-orderable==1.2.1
django-orderedmodel==2014.12.1
django-suit==0.2.12
easy-thumbnails==2.2
filebrowser-safe==0.3.8
Flask==0.10.1
Flask-WTF==0.11
future==0.14.3
gdata==2.0.18
grappelli-safe==0.3.13
html5lib==0.999
httplib2==0.9.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Mezzanine==3.1.10
oauth2client==1.4.7
oauthlib==0.7.2
pdb==0.1
Pillow==2.8.1
prettytable==0.7.2
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-gnupg==0.3.7
pytz==2015.2
PyYAML==3.11
requests==2.6.0
requests-oauthlib==0.4.2
rsa==3.1.4
six==1.9.0
tablib==0.10.0
tabulate==0.7.5
texttable==0.8.3
tzlocal==1.1.2
Werkzeug==0.10.4
WTForms==2.0.2
